Add skeleton of SupplementalProcessManagerService 1/n
Simply creates aidl file for the service and it's corresponding java
class in appropriate directory.

More skeleton code will be added on follow up CL. Uploaded this CL early
to unblock others who will be working on other parts of the skeleton
code.

Unit tests will be added when we start filling in the the
implmenetations.

Bug: 204991850
Test: m service-supplementalprocess
Change-Id: I33a69accccc73277d332f2926bde97b805f7089f