Replace stub implementation in SupplementalProcessManagerService
Previously, we had stub implementation that directly called
RemoteCodeCallback with successs when loadCode/requestSurfacePackage
was invoked.

We have now replaced it with call to SupplemenalProcessService and added
unit test to ensure that when SupplemenalProcessService calls back into
SupplementalProcessManagerService, the callback is handled properly.

We also created a two-way communication channel between manager
and remote code. Once code is loaded, app will try to
requestSurfacePackage using codeToken. We need to establish a way for
manager to call into the loaded code. This is done by having the
SupplementalProcess return a callback when it loads the code successfully.

The communication logic between manager and remote code in
SupplementalProcess is encapsulated in TwoWayCallback object.

The manager now commnicates with:
1) Remote code in SupplementalProcess using TwoWayCallback
2) App using RemoteCodeCallback object

Everything is glued together with codeToken.

It probably makes sense to combine all of codeToken, TwoWayCallback and
RemoteCodeCallback into one object since there is a one-to-one
relationship among all of them. To be done in follow up cl.

Bug: 204991850
Test: atest SupplementalProcessManagerServiceUnitTest
Change-Id: Iae4bea3f34fe3e0647a882de4d43fa3110b93a22