Load code provider within Supplemental Process
This change adds logic to load a code provider within the
supplemental process. The code to load is indicated by
the source directory of the passed ApplicationInfo. A
code provider app may implement the CodeProvider class, and
pass this class as a parameter when loading into the
supplemental process.

Added a test app and altered the unit tests to test this
functionality.

Bug: 204989872
Test: atest SupplementalProcessTests
Change-Id: If8f5b88baabb376c9c5180749b21c5a92c473e96