Add manual testing apps for SupplementalProcess
This change introduces two test apps: SupplementalProcessClient
and SupplementalProcessCodeProvider. The former loads the latter
into the Supplemental Process, and the code provider generates
a view which is rendered inside the client app.

Also added some logic for passing codeProviderClassName from the
client app (this is an interim solution), as well as passing the
width and height of the view that should be rendered.

Test: Manual. Build and install both packages and verify
      the loading and rendering behavior.
Bug: 208982128
Change-Id: Ic240b9a15ff4199e9b21022888bb046f34512fc6