Dexopt service-supplementalprocess.
`service-supplementalprocess` is a jar that system_server loads
dynamically. Such jars can be dexopted now. This CL adds
`service-supplementalprocess` to a `systemserverclasspath_fragment`,
which does not actually add the jar to the classpath, but tells the
build system to dexopt it.

NOTE: The corresponding `systemserverclasspath_fragment` needs to be
added to the `sdk` once the SDK is created, and
`exported_systemserverclasspath_fragments` needs to be added to the
`prebuilt_apex` once a new prebuilt is dropped.

Bug: 203198541
Test: m com.android.supplementalprocess
Change-Id: Ib689486528a3f1c9bf0f7936ca0420bc14e44fa0