Decouple LinkToDeath subscription from ServiceProvider
ServiceProvider used to unregister App when it is get killed by
subscribing to linkToDeath event,
in the current design we keep binding the service automatically,
ServiceProvider then keeps subscribing to the same event multiple times.

This CL is decoupling this subscription from ServiceProvider by moving
it totally away from ServiceProvider to the ServiceManager.

This CL also:
- Remove the dependency over the appCallback binder from ServiceProvider
(It is totally unrelated to the provider)
- Adding unregisterApp to the provider interface, to be called by the
manager when linkToDeath happens
- Clean some parts of code by reordering fields and methods by access
modifiers

Bug: 209006687
Test: asest testSupplementalProcessUnbindingWhenAppDied
Change-Id: I4303ab9128cc0e899a07b9819a7f207a100e9457