Create a new supplemental process service per app
Using the newly proposed API, a new supplemental process service is
created for each uid that asks to load code. Each instance is uniquely
identified by an instance name created using the calling uid.

This CL completely replaces the old implementation where there was a
supplemental process for each user. Uids and the corresponding
supplemental process are kept track of. Like before, each supplemental
process is non-persistent and dependent on the life of the app.
Unbinding of the supplemental process now occurs directly when the app
dies as opposed to before when registerApp() was used to keep track of
the apps hosted by the supplemental process.

Since the API is only callable by the system server, the unit test
calling it fails and is removed.

Bug: 209058402
Bug: 210570418
Bug: 203670795
Test: Manual, installing two client apps and attempting to load code
from each and examining new processes created. Hostside tests are WIP.

Change-Id: I6053440ef30439a752b972b1e917333d30a98c48