Add CodeContext wrapper for accessing resources
This change introduces a new CodeContext class, which will be
available to CodeProvider classes. The primary function of this
wrapper is to handle calls from the code provider to access its
own resources.

Also adds resources to the manual testing apps as well as the
end to end tests to ensure that both resources and assets work
with the CodeContext. Adds unit tests to check that
CodeContext.getResources() works correctly.

Test: Manually load and test apps
Test: atest SupplementalProcessManagerTest
Test: atest SupplementalProcessManagerServiceUnitTest
Bug: 217348079
Change-Id: I20105f6a0df9c6e1b97ea125f86a9c51fbefae4c