SupplementalProcess: remove getUid and use toSdkSandboxUid
As Process.toSdkSandboxUid is supported, we no longer need the getUid
command. This allows us to make the ISdkSandboxService oneway
which fixes an error from Binder.
This change required updating a unit test which depended on a -1 being
returned from getUid. Now we mock the ActivityManager::killUid method
using a spy to avoid throwing a SecurityException during a test.

Fixes: 219962323
Test: adb logcat | grep "FLAG_ONEWAY" (see no output)
Change-Id: Ib30a575d9145f4f8720fcac9b52d3677208fdffe