Add intent checking to SdkSandboxManagerLocal
Adds two system server in-process APIs to check whether a given
intent may be broadcast or an activity may be started from the sdk
sandbox process. This will be called from ActivityManager when trying
to send a broadcast or start an activity from a uid in the sdk sandbox range.

Bug: 209599396
Test: atest SdkSandboxManagerServiceUnitTests
Test: Manual. Try to broadcast intent using manual test app.
Change-Id: I9ea7809b22f5e7d3bf10f53a5f3264e6c8f53377